AppCode/requirements.txt
Flask>=2.0
flask-cors
google-cloud-aiplatform
google-cloud-firestore
vertexai
gunicorn